feat: subjectInfo

+++ jc-club-subject/jc-club-infra/src/main/java/com/jingdianjichi/subject/infra/basic/entity/SubjectMultiple.java
@@ -25,7 +25,7 @@ public class SubjectMultiple implements Serializable {
 /**
      * 选项类型
      */
-    private Long optionType;
+    private Integer optionType;
 /**
      * 选项内容
      */
